+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 ### Added
 - Create, update and delete app stores
 - Settings: Radio buttons to use/not use proxy
+- Settings: Show available DSC update
 
 ### Fixes
 - Settings: Correct handling of proxy username/password (new DSC attribute to detect set authentication)